+++ unit-tests/json-to-iac/config-to-files-json.test.js
@@ -208,6 +208,12 @@ variable "prefix" {
   }
 }
 
+variable "account_id" {
+  description = "IBM Account ID where resources will be provisioned"
+  type        = string
+  default     = "1234"
+}
+
 variable "slz_ssh_key_public_key" {
   description = "Public SSH Key Value for Slz SSH Key"
   type        = string
@@ -350,7 +356,6 @@ variable "secrets_manager_imported_cert_data" {
           name: "event-streams",
           plan: "enterprise-3nodes-2tb",
           resource_group: "slz-service-rg",
-          endpoints: "private",
           private_ip_allowlist: ["10.0.0.0/32", "10.0.0.1/32"],
           throughput: "150MB/s",
           storage_size: "2TB",
@@ -370,7 +375,6 @@ variable "secrets_manager_imported_cert_data" {
           logdna: {
             enabled: true,
             plan: "lite",
-            endpoints: "private",
             platform_logs: true,
             resource_group: "service-rg",
             role: "Manager",
@@ -382,7 +386,6 @@ variable "secrets_manager_imported_cert_data" {
           sysdig: {
             enabled: true,
             plan: "lite",
-            endpoints: "private",
             resource_group: "service-rg",
             platform_logs: true,
           },
@@ -394,7 +397,6 @@ variable "secrets_manager_imported_cert_data" {
             bucket: "atracker",
             cos_key: "atracker-cos-key",
             plan: "lite",
-            endpoints: "private",
             resource_group: "service-rg",
             archive: true,
             instance: true,
@@ -573,7 +575,6 @@ terraform {
           logdna: {
             enabled: true,
             plan: "lite",
-            endpoints: "private",
             platform_logs: true,
             resource_group: "service-rg",
             role: "Manager",
@@ -585,7 +586,6 @@ terraform {
           sysdig: {
             enabled: true,
             plan: "lite",
-            endpoints: "private",
             resource_group: "service-rg",
             platform_logs: true,
           },
@@ -597,7 +597,6 @@ terraform {
             bucket: "atracker",
             cos_key: "atracker-cos-key",
             plan: "lite",
-            endpoints: "private",
             resource_group: "service-rg",
             archive: false,
             instance: true,
@@ -1000,6 +999,11 @@ variable "prefix" {
     error_message = "Prefix must begin with a lowercase letter and contain only lowercase letters, numbers, and - characters. Prefixes must end with a lowercase letter or number and be 16 or fewer characters."
     condition     = can(regex("^([a-z]|[a-z][-a-z0-9]*[a-z0-9])", var.prefix)) && length(var.prefix) <= 16
   }
+}
+
+variable "account_id" {
+  description = "IBM Account ID where resources will be provisioned"
+  type        = string
 }
 
 variable "slz_ssh_key_public_key" {